优化代码清晰度和运算逻辑
在多个测试类中，移除了对`Radians`属性的强制类型转换，直接使用其进行操作，表明`angle.Radians`已经是`float`类型或相关方法已支持其实际类型。同时，在数学和几何计算中增加了括号来明确运算的优先级，提高了代码的可读性并减少了因优先级混淆可能导致的错误。在`AssimpParsing.cs`文件中，简化了对`MTextureCoords[0][i]`的引用，去除了不必要的间接引用操作。此外，通过增加括号明确了乘法和加法的执行顺序，以避免潜在的逻辑错误，并在多个结构体和类中改进了向量和矩阵的运算，使得运算逻辑更加清晰，可能也提高了运算效率。最后，通过调整计算表达式，使得光照和颜色的计算在渲染器的实现中更加直观和清晰。这些改动旨在提高代码的清晰度、准确性和效率，同时也减少了对特定类型转换的依赖，使得代码更加健壮。

+++ Mathematics/Maths.Tests/Matrix3X3Tests.cs
@@ -18,7 +18,7 @@ public class Matrix3X3Tests
         Vector2 point2 = new(1, 1);
 
         Matrix3x3d matrix1 = Matrix3x3d.CreateRotation(angle);
-        Matrix3x2 matrix2 = Matrix3x2.CreateRotation((float)angle.Radians);
+        Matrix3x2 matrix2 = Matrix3x2.CreateRotation(angle.Radians);
 
         point1 = matrix1 * point1;
         point2 = Vector2.Transform(point2, matrix2);
@@ -88,7 +88,7 @@ public class Matrix3X3Tests
         Matrix3x3d r1 = Matrix3x3d.CreateRotation(angle);
         Matrix3x3d s1 = Matrix3x3d.CreateScale(new Vector2d(x, y));
         Matrix3x3d t1 = Matrix3x3d.CreateTranslation(new Vector2d(x, y));
-        Matrix3x2 r2 = Matrix3x2.CreateRotation((float)angle.Radians);
+        Matrix3x2 r2 = Matrix3x2.CreateRotation(angle.Radians);
         Matrix3x2 s2 = Matrix3x2.CreateScale(new Vector2((float)x, (float)y));
         Matrix3x2 t2 = Matrix3x2.CreateTranslation(new Vector2((float)x, (float)y));
